requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
pydantic_ai
httpx
python-dotenv
prisma
sqlalchemy
PyJWT
